safari-600.7-branch/LayoutTests/fast/dom/custom/lifecycle-ready-createElement-recursion-expected.txt

CONSOLE MESSAGE: line 9: Uncaught RangeError: Maximum call stack size exceeded.
PASS unless crash
PASS successfullyParsed is true

TEST COMPLETE
